Generalize and introduce Sealer into CP.
Sealer is a special consumer that provides information regarding where the cp is up to.
It will be the first one during cp switch over , as a conservative marker of everything
before or equals to this point, should be in current cp, possibly some consumer are above this point which is fine.
And Sealer is the last one during cp flush after all other services flushed successfully.

Signed-off-by: Xiaoxi Chen <xiaoxchen@ebay.com>