Support Baseline Resync (#596)
* Support Baseline resync

For Nuraft baseline resync, we separate the process into two layers: HomeStore layer and Application layer.
We use the first bit of the obj_id to indicate the message type: 0 is for HS, 1 is for Application.

In the HomeStore layer, leader needs to transmit the DSN to the follower, this is intended to handle the following case:

1. Leader sends snapshot at LSN T1 to follower F1.
2. F1 fully receives the snapshot and now at T1.
3. Leader yield its leadership, F1 elected as leader.
In this sequence the incremental resync will not kicked in to update the m_next_dsn, and as result, duplication may occur.